+++ src/routers/api/sessions.api.js
@@ -2,6 +2,12 @@ import passport from "../../middlewares/passport.mid.js"; //Dejar
 import passportCb from "../../middlewares/passportCb.mid.js";
 import { isValidData } from "../../middlewares/isValidData.mid.js";
 import CustomRouter from "../CustomRouter.js";
+import {
+  create,
+  login,
+  online,
+  signout,
+} from "../../controllers/sessions.controller.js";
 
 class SessionsRouter extends CustomRouter {
   init() {
@@ -17,53 +23,5 @@ class SessionsRouter extends CustomRouter {
     this.create("/signout", ["USER", "ADMIN"], passportCb("jwt"), signout);
   }
 }
-
-//Create a new user
-const create = async (req, res, next) => {
-  try {
-    return res.message201("User created successfully");
-  } catch (error) {
-    return next(error);
-  }
-};
-
-//login user
-const login = (req, res, next) => {
-  try {
-    return res
-      .cookie("token", req.user.token, { signedCookie: true })
-      .message200("Login successful");
-  } catch (error) {
-    return next(error);
-  }
-};
-
-//Online User
-const online = async (req, res, next) => {
-  try {
-    if (req.user.online) {
-      const user_online = {
-        email: req.user.email,
-        role: req.user.role,
-        photo: req.user.photo,
-        user_id: req.user._id,
-      };
-      res.response200(user_online);
-    }
-  } catch (error) {
-    return next(error);
-  }
-};
-
-//login user
-const signout = async (req, res, next) => {
-  try {
-    return req.user.online
-      ? res.clearCookie("token").message200("Signed out!")
-      : res.error404("Invalid credentials from signout!");
-  } catch (error) {
-    return next(error);
-  }
-};
 
 export const sessionsRouter = new SessionsRouter().getRouter();
